apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
psycopg[binary]==3.2.9
redis==6.4.0
prometheus-client==0.22.1
sqlalchemy==2.0.41
alembic==1.16.4
PyJWT==2.10.1
pwdlib==0.2.1
argon2-cffi==25.1.0
pydantic-settings==2.10.1
email-validator==2.2.0
pytest==8.4.1
ruff==0.12.0
httpx==0.28.1
